refactor: move input/output args to standalone file
